Removed unused class, irrelevant tests
- Removed the TokenResponse class. The feature/TM/EmailRoutes branch
was created from feature/OAuth2 branch where the TokenResponse class
was removed after creating the feature/TM/EmailRoutes class, so we
removed it here also since it is not used anymore.
- We removed the RuntimeInformationProvider.spec.js file on the parent
branch since the format of the User-Agent string changed. Since
feature/TM/EmailRoutes was created before removing the file, we had to
remove it here also
- Removed the timeouts from all the specs
- Removed unused import in the BaseApi class